feat: 优化对话入口方法代码和快速对话接口
