+++ packages/derisk-serve/src/derisk_serve/agent/evaluation/evaluation.py
@@ -39,7 +39,7 @@ class AgentOutputOperator(MapOperator):
         final_output = None
         try:
             begin_time_ms = int(datetime.now().timestamp())
-            async for output in multi_agents.app_chat(
+            async for output,agent_conv_id in multi_agents.app_chat(
                 conv_uid=str(uuid.uuid1()),
                 gpts_name=self.app_code,
                 user_query=input_value,
